# docs/swagger.yaml
definitions:
  controller.ResponseHTTP:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/monitor:
    get:
      consumes:
      - application/json
      description: 获取监控数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取监控数据
      tags:
      - 获取监控数据
    post:
      consumes:
      - application/json
      description: 添加监控数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 添加监控数据
      tags:
      - 添加监控数据
  /v1/region:
    get:
      consumes:
      - application/json
      description: 获取省市区
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 获取省市区
      tags:
      - 获取省市区
  /v1/test:
    get:
      consumes:
      - application/json
      description: Test
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: Test
      tags:
      - Test
  /v1/upload:
    post:
      consumes:
      - application/json
      description: 文件上传
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ResponseHTTP'
      summary: 文件上传
      tags:
      - 文件上传
swagger: "2.0"